hw items 3

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -3,11 +3,12 @@ package ru.javawebinar.topjava.repository.inmemory;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.web.SecurityUtil;
 
-import java.util.Collection;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 public class InMemoryMealRepository implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
@@ -30,17 +31,28 @@ public class InMemoryMealRepository implements MealRepository {
 
     @Override
     public boolean delete(int id) {
-        return repository.remove(id) != null;
+        if (repository.get(id).getUserId() == SecurityUtil.authUserId()) {
+            return repository.remove(id) != null;
+        }
+        else return false;
     }
 
     @Override
     public Meal get(int id) {
-        return repository.get(id);
+        if (repository.get(id).getUserId() == SecurityUtil.authUserId()) {
+            return repository.get(id);
+        }
+        else return null;
     }
 
     @Override
     public Collection<Meal> getAll() {
-        return repository.values();
+
+       return  repository.values().stream().filter(meal -> meal.getUserId()==SecurityUtil.authUserId()).
+               sorted((o1, o2)-> o2.getDateTime().compareTo(o1.getDateTime())).collect(Collectors.toCollection(ArrayList::new));
+
+
+
     }
 }
